Separate WCAGColorSuggestions, WCAGComplianceResult, WCAGContrastLevel into their own files
