+++ Sources/ColorKit/WCAG/WCAGCompliance.swift
@@ -1,88 +1,5 @@
 import SwiftUI
 
-/// Represents the WCAG contrast ratio levels
-public enum WCAGContrastLevel: String, CaseIterable, Identifiable, Sendable {
-    case AALarge = "AA Large"
-    case AA = "AA"
-    case AAALarge = "AAA Large"
-    case AAA = "AAA"
-
-    public var id: String { rawValue }
-
-    /// The minimum contrast ratio required for this level
-    public var minimumRatio: Double {
-        switch self {
-        case .AALarge:
-            return 3.0
-        case .AA:
-            return 4.5
-        case .AAALarge:
-            return 4.5
-        case .AAA:
-            return 7.0
-        }
-    }
-
-    /// Description of the WCAG level
-    public var description: String {
-        switch self {
-        case .AALarge:
-            return "AA level for large text (18pt+)"
-        case .AA:
-            return "AA level for normal text"
-        case .AAALarge:
-            return "AAA level for large text (18pt+)"
-        case .AAA:
-            return "AAA level for normal text"
-        }
-    }
-}
-
-/// Represents the result of a WCAG compliance check between two colors.
-/// This struct stores the contrast ratio and whether the colors pass various WCAG compliance levels.
-public struct WCAGComplianceResult {
-    /// The calculated contrast ratio between two colors.
-    public let contrastRatio: Double
-
-    /// Whether the colors pass the AA compliance level for normal text.
-    public let passesAA: Bool
-
-    /// Whether the colors pass the AA compliance level for large text (18pt+).
-    public let passesAALarge: Bool
-
-    /// Whether the colors pass the AAA compliance level for normal text.
-    public let passesAAA: Bool
-
-    /// Whether the colors pass the AAA compliance level for large text (18pt+).
-    public let passesAAALarge: Bool
-
-    /// Returns the highest WCAG compliance level that the colors pass.
-    /// If none are met, it returns `nil`.
-    public var highestLevel: WCAGContrastLevel? {
-        if passesAAA {
-            return .AAA
-        } else if passesAAALarge {
-            return .AAALarge
-        } else if passesAA {
-            return .AA
-        } else if passesAALarge {
-            return .AALarge
-        } else {
-            return nil
-        }
-    }
-
-    /// Returns a list of all WCAG compliance levels that the colors pass.
-    public var passes: [WCAGContrastLevel] {
-        var result: [WCAGContrastLevel] = []
-        if passesAALarge { result.append(.AALarge) }
-        if passesAA { result.append(.AA) }
-        if passesAAALarge { result.append(.AAALarge) }
-        if passesAAA { result.append(.AAA) }
-        return result
-    }
-}
-
 // Extension to get RGBA components
 extension Color {
     /// Get the RGBA components of a color
@@ -179,92 +96,6 @@ public extension Color {
             // Suggest a dark color
             return .black
         }
-    }
-}
-
-/// Provides advanced color suggestions to achieve WCAG compliance
-public struct WCAGColorSuggestions {
-    private let baseColor: Color
-    private let targetColor: Color
-    private let targetLevel: WCAGContrastLevel
-
-    /// Creates a new WCAG color suggestions generator
-    /// - Parameters:
-    ///   - baseColor: The base color that won't be changed
-    ///   - targetColor: The color that needs to be adjusted for compliance
-    ///   - targetLevel: The WCAG compliance level to aim for
-    public init(baseColor: Color, targetColor: Color, targetLevel: WCAGContrastLevel = .AA) {
-        self.baseColor = baseColor
-        self.targetColor = targetColor
-        self.targetLevel = targetLevel
-    }
-
-    /// Generates a set of color suggestions that meet WCAG compliance
-    /// - Parameter preserveHue: Whether to preserve the original hue when generating suggestions
-    /// - Returns: An array of suggested colors that meet the required contrast level
-    public func generateSuggestions(preserveHue: Bool = true) -> [Color] {
-        let currentCompliance = baseColor.wcagCompliance(with: targetColor)
-
-        // If already compliant, return the original color
-        if currentCompliance.passes.contains(targetLevel) {
-            return [targetColor]
-        }
-
-        var suggestions: [Color] = []
-
-        // Get target color components
-        guard let hsl = targetColor.hslComponents() else {
-            return [targetColor]
-        }
-
-        let baseLuminance = baseColor.wcagRelativeLuminance()
-
-        // Determine if we need to lighten or darken
-        let needsDarkening = baseLuminance > 0.5
-
-        // Generate a lighter/darker version maintaining hue and saturation
-        if preserveHue {
-            // Adjust lightness in increments
-            let steps = 10
-            for i in 1...steps {
-                let adjustedLightness = needsDarkening
-                    ? max(0.0, hsl.lightness - (Double(i) / Double(steps)) * hsl.lightness)
-                    : min(1.0, hsl.lightness + (Double(i) / Double(steps)) * (1.0 - hsl.lightness))
-
-                if let adjustedColor = Color.fromHSL(hue: hsl.hue, saturation: hsl.saturation, lightness: CGFloat(adjustedLightness)) {
-                    let compliance = baseColor.wcagCompliance(with: adjustedColor)
-                    if compliance.passes.contains(targetLevel) {
-                        suggestions.append(adjustedColor)
-                        break
-                    }
-                }
-            }
-        }
-
-        // If we couldn't find a suitable color by adjusting lightness alone,
-        // try adjusting saturation as well
-        if suggestions.isEmpty {
-            for saturationFactor in stride(from: 0.8, to: 0.0, by: -0.2) {
-                let adjustedSaturation = hsl.saturation * saturationFactor
-
-                let adjustedLightness = needsDarkening ? 0.0 : 1.0
-
-                if let adjustedColor = Color.fromHSL(hue: hsl.hue, saturation: adjustedSaturation, lightness: CGFloat(adjustedLightness)) {
-                    let compliance = baseColor.wcagCompliance(with: adjustedColor)
-                    if compliance.passes.contains(targetLevel) {
-                        suggestions.append(adjustedColor)
-                        break
-                    }
-                }
-            }
-        }
-
-        // If still no suggestions, add fallback colors
-        if suggestions.isEmpty {
-            suggestions.append(needsDarkening ? .black : .white)
-        }
-
-        return suggestions
     }
 }
